Added Repositories

+++ src/main/java/com/trading212/code212/repositories/ProblemRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface ProblemRepository {
 
     ProblemEntity createProblem(String title, String description, String inputUrl, String outputUrl);
-    Optional<ProblemEntity> getProblemById(int id);
-    void deleteProblemById(int id);
+    Optional<ProblemEntity> getProblemById(Long id);
+    void deleteProblemById(Long id);
 
 }
